requirements.txt
discord.py===1.3.3
PyYAML~=5.1